+++ server/main.go
@@ -32,7 +32,7 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-var blocked = cmap.New()
+var blocked = cmap.New[int64]()
 var lastRequest = time.Now().Unix()
 
 func main() {
@@ -134,9 +134,8 @@ func wsHandshake(ctx *gin.Context) {
 		return
 	}
 	secret := append(utils.GetUUID(), utils.GetUUID()...)
-	err = common.Melody.HandleRequestWithKeys(ctx.Writer, ctx.Request, http.Header{
-		`Secret`: []string{hex.EncodeToString(secret)},
-	}, gin.H{
+	ctx.Writer.Header().Add(`Secret`, hex.EncodeToString(secret))
+	err = common.Melody.HandleRequestWithKeys(ctx.Writer, ctx.Request, gin.H{
 		`Secret`:   secret,
 		`LastPack`: utils.Unix,
 		`Address`:  common.GetRemoteAddr(ctx),
@@ -158,20 +157,38 @@ func wsOnMessage(session *melody.Session, _ []byte) {
 func wsOnMessageBinary(session *melody.Session, data []byte) {
 	var pack modules.Packet
 
-	{
-		dataLen := len(data)
-		if dataLen >= 22 {
-			if bytes.Equal(data[:5], []byte{34, 22, 19, 17, 20}) {
-				event := hex.EncodeToString(data[6:22])
-				copy(data[6:], data[22:])
-				common.CallEvent(modules.Packet{
-					Event: event,
-					Data: gin.H{
-						`data`: utils.GetSlicePrefix(&data, dataLen-16),
-					},
-				}, session)
-				return
+	dataLen := len(data)
+	if dataLen > 24 {
+		if service, op, isBinary := utils.CheckBinaryPack(data); isBinary {
+			switch service {
+			case 20:
+				switch op {
+				case 00, 01, 02, 03:
+					event := hex.EncodeToString(data[6:22])
+					copy(data[6:], data[22:])
+					common.CallEvent(modules.Packet{
+						Act:   `RAW_DATA_ARRIVE`,
+						Event: event,
+						Data: gin.H{
+							`data`: utils.GetSlicePrefix(&data, dataLen-16),
+						},
+					}, session)
+				}
+			case 21:
+				switch op {
+				case 00, 01:
+					event := hex.EncodeToString(data[6:22])
+					copy(data[6:], data[22:])
+					common.CallEvent(modules.Packet{
+						Act:   `RAW_DATA_ARRIVE`,
+						Event: event,
+						Data: gin.H{
+							`data`: utils.GetSlicePrefix(&data, dataLen-16),
+						},
+					}, session)
+				}
 			}
+			return
 		}
 	}
 
@@ -195,14 +212,13 @@ func wsOnMessageBinary(session *melody.Session, data []byte) {
 }
 
 func wsOnDisconnect(session *melody.Session) {
-	if val, ok := common.Devices.Get(session.UUID); ok {
-		deviceInfo := val.(*modules.Device)
-		terminal.CloseSessionsByDevice(deviceInfo.ID)
-		desktop.CloseSessionsByDevice(deviceInfo.ID)
+	if device, ok := common.Devices.Get(session.UUID); ok {
+		terminal.CloseSessionsByDevice(device.ID)
+		desktop.CloseSessionsByDevice(device.ID)
 		common.Info(nil, `CLIENT_OFFLINE`, ``, ``, map[string]any{
 			`device`: map[string]any{
-				`name`: deviceInfo.Hostname,
-				`ip`:   deviceInfo.WAN,
+				`name`: device.Hostname,
+				`ip`:   device.WAN,
 			},
 		})
 		// save devices if offline
@@ -275,10 +291,9 @@ func pingDevice(s *melody.Session) {
 	trigger := utils.GetStrUUID()
 	common.SendPack(modules.Packet{Act: `PING`, Event: trigger}, s)
 	common.AddEventOnce(func(packet modules.Packet, session *melody.Session) {
-		val, ok := common.Devices.Get(s.UUID)
+		device, ok := common.Devices.Get(s.UUID)
 		if ok {
-			deviceInfo := val.(*modules.Device)
-			deviceInfo.Latency = uint(time.Now().UnixMilli()-t) / 2
+			device.Latency = uint(time.Now().UnixMilli()-t) / 2
 		}
 	}, s.UUID, trigger, 3*time.Second)
 }
@@ -286,12 +301,12 @@ func pingDevice(s *melody.Session) {
 func checkAuth() gin.HandlerFunc {
 	// Token as key and update timestamp as value.
 	// Stores authenticated tokens.
-	tokens := cmap.New()
+	tokens := cmap.New[int64]()
 	go func() {
 		for now := range time.NewTicker(60 * time.Second).C {
 			var queue []string
-			tokens.IterCb(func(key string, v any) bool {
-				if now.Unix()-v.(int64) > 1800 {
+			tokens.IterCb(func(key string, t int64) bool {
+				if now.Unix()-t > 1800 {
 					queue = append(queue, key)
 				}
 				return true
@@ -299,8 +314,8 @@ func checkAuth() gin.HandlerFunc {
 			tokens.Remove(queue...)
 			queue = nil
 
-			blocked.IterCb(func(addr string, v any) bool {
-				if now.Unix() > v.(int64) {
+			blocked.IterCb(func(addr string, t int64) bool {
+				if now.Unix() > t {
 					queue = append(queue, addr)
 				}
 				return true
@@ -333,7 +348,7 @@ func checkAuth() gin.HandlerFunc {
 		if !passed {
 			addr := common.GetRealIP(ctx)
 			if expire, ok := blocked.Get(addr); ok {
-				if now < expire.(int64) {
+				if now < expire {
 					ctx.AbortWithStatusJSON(http.StatusTooManyRequests, modules.Packet{Code: 1})
 					return
 				}
@@ -345,7 +360,7 @@ func checkAuth() gin.HandlerFunc {
 
 			if ctx.IsAborted() {
 				blocked.Set(addr, now+1)
-				user = utils.If(len(user) == 0, `EMPTY`, user)
+				user = utils.If(len(user) == 0, `<EMPTY>`, user)
 				common.Warn(ctx, `LOGIN_ATTEMPT`, `fail`, ``, map[string]any{
 					`user`: user,
 				})