+++ web/src/components/desktop/desktop.jsx
@@ -1,8 +1,7 @@
 import React, {useCallback, useEffect, useState} from 'react';
-import {encrypt, decrypt, formatSize, genRandHex, getBaseURL, translate} from "../utils/utils";
-import i18n from "../locale/locale";
-import DraggableModal from "./modal";
-import CryptoJS from "crypto-js";
+import {encrypt, decrypt, formatSize, genRandHex, getBaseURL, translate, str2ua, hex2ua, ua2hex} from "../../utils/utils";
+import i18n from "../../locale/locale";
+import DraggableModal from "../modal";
 import {Button, message} from "antd";
 import {FullscreenOutlined, ReloadOutlined} from "@ant-design/icons";
 
@@ -17,37 +16,39 @@ let bytes = 0;
 let ticks = 0;
 let title = i18n.t('DESKTOP.TITLE');
 function ScreenModal(props) {
+	const [resolution, setResolution] = useState('0x0');
 	const [bandwidth, setBandwidth] = useState(0);
 	const [fps, setFps] = useState(0);
 	const canvasRef = useCallback((e) => {
-		if (e && props.visible && !conn) {
+		if (e && props.open && !conn && !canvas) {
+			secret = hex2ua(genRandHex(32));
 			canvas = e;
 			initCanvas(canvas);
 			construct(canvas);
 		}
 	}, [props]);
 	useEffect(() => {
-		if (props.visible) {
-			secret = CryptoJS.enc.Hex.parse(genRandHex(32));
-		} else {
+		if (!props.open) {
+			canvas = null;
 			if (ws && conn) {
 				clearInterval(ticker);
 				ws.close();
 				conn = false;
 			}
 		}
-	}, [props.visible, props.device]);
+	}, [props.open]);
 
 	function initCanvas() {
 		if (!canvas) return;
 		ctx = canvas.getContext('2d', {alpha: false});
+		ctx.imageSmoothingEnabled = false;
 	}
 	function construct() {
 		if (ctx !== null) {
 			if (ws !== null && conn) {
 				ws.close();
 			}
-			ws = new WebSocket(getBaseURL(true, `api/device/desktop?device=${props.device.id}&secret=${secret}`));
+			ws = new WebSocket(getBaseURL(true, `api/device/desktop?device=${props.device.id}&secret=${ua2hex(secret)}`));
 			ws.binaryType = 'arraybuffer';
 			ws.onopen = () => {
 				conn = true;
@@ -79,38 +80,25 @@ function ScreenModal(props) {
 				ticks++;
 				if (ticks > 10 && conn) {
 					ticks = 0;
-					ws.send(encrypt({
+					sendData({
 						act: 'DESKTOP_PING'
-					}, secret));
+					});
 				}
 			}, 1000);
 		}
 	}
 	function fullScreen() {
-		try {
-			canvas.requestFullscreen();
-		} catch {}
-		try {
-			canvas.webkitRequestFullscreen();
-		} catch {}
-		try {
-			canvas.mozRequestFullScreen();
-		} catch {}
-		try {
-			canvas.msRequestFullscreen();
-		} catch {}
+		canvas.requestFullscreen().catch(console.error);
 	}
 	function refresh() {
-		if (canvas && props.visible) {
+		if (canvas && props.open) {
 			if (!conn) {
-				canvas.width = 1920;
-				canvas.height = 1080;
 				initCanvas(canvas);
 				construct(canvas);
 			} else {
-				ws.send(encrypt({
+				sendData({
 					act: 'DESKTOP_SHOT'
-				}, secret));
+				});
 			}
 		}
 	}
@@ -124,23 +112,25 @@ function ScreenModal(props) {
 			return;
 		}
 		if (op === 2) {
-			let width = dv.getUint16(1, false);
-			let height = dv.getUint16(3, false);
+			let width = dv.getUint16(3, false);
+			let height = dv.getUint16(5, false);
 			if (width === 0 || height === 0) return;
 			canvas.width = width;
 			canvas.height = height;
+			setResolution(`${width}x${height}`);
 			return;
 		}
 		if (op === 0) frames++;
 		bytes += ab.byteLength;
 		let offset = 1;
 		while (offset < ab.byteLength) {
-			let it = dv.getUint16(offset + 0, false); // image type
-			let il = dv.getUint16(offset + 2, false); // image length
+			let bl = dv.getUint16(offset + 0, false); // body length
+			let it = dv.getUint16(offset + 2, false); // image type
 			let dx = dv.getUint16(offset + 4, false); // image block x
 			let dy = dv.getUint16(offset + 6, false); // image block y
 			let bw = dv.getUint16(offset + 8, false); // image block width
 			let bh = dv.getUint16(offset + 10, false); // image block height
+			let il = bl - 10; // image length
 			offset += 12;
 			updateImage(ab.slice(offset, offset + il), it, dx, dy, bw, bh, canvasCtx);
 			offset += il;
@@ -148,13 +138,18 @@ function ScreenModal(props) {
 		dv = null;
 	}
 	function updateImage(ab, it, dx, dy, bw, bh, canvasCtx) {
-		if (it === 0) {
-			canvasCtx.putImageData(new ImageData(new Uint8ClampedArray(ab), bw, bh), dx, dy, 0, 0, bw, bh);
-		} else {
-			createImageBitmap(new Blob([ab]), 0, 0, bw, bh)
-			.then((ib) => {
-				canvasCtx.drawImage(ib, 0, 0, bw, bh, dx, dy, bw, bh);
-			});
+		switch (it) {
+			case 0:
+				canvasCtx.putImageData(new ImageData(new Uint8ClampedArray(ab), bw, bh), dx, dy, 0, 0, bw, bh);
+				break;
+			case 1:
+				createImageBitmap(new Blob([ab]), 0, 0, bw, bh, {
+					premultiplyAlpha: 'none',
+					colorSpaceConversion: 'none'
+				}).then((ib) => {
+					canvasCtx.drawImage(ib, 0, 0, bw, bh, dx, dy, bw, bh);
+				});
+				break;
 		}
 	}
 	function handleJSON(ab) {
@@ -173,12 +168,23 @@ function ScreenModal(props) {
 		}
 	}
 
+	function sendData(data) {
+		if (conn) {
+			let body = encrypt(str2ua(JSON.stringify(data)), secret);
+			let buffer = new Uint8Array(body.length + 8);
+			buffer.set(new Uint8Array([34, 22, 19, 17, 20, 3]), 0);
+			buffer.set(new Uint8Array([body.length >> 8, body.length & 0xFF]), 6);
+			buffer.set(body, 8);
+			ws.send(buffer);
+		}
+	}
+
 	return (
 		<DraggableModal
 			draggable={true}
 			maskClosable={false}
 			destroyOnClose={true}
-			modalTitle={`${title} ${formatSize(bandwidth)}/s FPS: ${fps}`}
+			modalTitle={`${title} ${resolution} ${formatSize(bandwidth)}/s FPS: ${fps}`}
 			footer={null}
 			height={480}
 			width={940}