+++ web/src/pages/overview.jsx
@@ -1,50 +1,31 @@
 import React, {useEffect, useRef, useState} from 'react';
 import ProTable, {TableDropdown} from '@ant-design/pro-table';
 import {Button, Image, message, Modal, Progress, Tooltip} from 'antd';
-import {catchBlobReq, formatSize, request, translate, tsToTime, waitTime} from "../utils/utils";
-import Generate from "../components/generate";
-import Explorer from "../components/explorer";
-import Terminal from "../components/terminal";
-import ProcMgr from "../components/procmgr";
-import Desktop from "../components/desktop";
-import Runner from "../components/runner";
+import {catchBlobReq, formatSize, request, tsToTime, waitTime} from "../utils/utils";
 import {QuestionCircleOutlined} from "@ant-design/icons";
 import i18n from "../locale/locale";
 
 // DO NOT EDIT OR DELETE THIS COPYRIGHT MESSAGE.
 console.log("%c By XZB %c https://github.com/XZB-1248/Spark", 'font-family:"Helvetica Neue",Helvetica,Arial,sans-serif;font-size:64px;color:#00bbee;-webkit-text-fill-color:#00bbee;-webkit-text-stroke:1px#00bbee;', 'font-size:12px;');
 
-function UsageBar(props) {
-	let {usage} = props;
-	usage = usage || 0;
-	usage = Math.round(usage * 100) / 100;
-
-	return (
-		<Tooltip
-			title={props.title??`${usage}%`}
-			overlayInnerStyle={{
-				whiteSpace: 'nowrap',
-				wordBreak: 'keep-all',
-				maxWidth: '300px',
-			}}
-			overlayStyle={{
-				maxWidth: '300px',
-			}}
-		>
-			<Progress percent={usage} showInfo={false} strokeWidth={12} trailColor='#FFECFF'/>
-		</Tooltip>
-	);
-}
+let ComponentMap = {
+	Generate: null,
+	Explorer: null,
+	Terminal: null,
+	ProcMgr: null,
+	Desktop: null,
+	Execute: null,
+};
 
 function overview(props) {
-	const [runner, setRunner] = useState(false);
+	const [loading, setLoading] = useState(false);
+	const [execute, setExecute] = useState(false);
 	const [desktop, setDesktop] = useState(false);
 	const [procMgr, setProcMgr] = useState(false);
 	const [explorer, setExplorer] = useState(false);
 	const [generate, setGenerate] = useState(false);
 	const [terminal, setTerminal] = useState(false);
 	const [screenBlob, setScreenBlob] = useState('');
-	const [isWindows, setIsWindows] = useState(false);
 	const [dataSource, setDataSource] = useState([]);
 	const [columnsState, setColumnsState] = useState(getInitColumnsState());
 
@@ -184,16 +165,34 @@ function overview(props) {
 		setting: true,
 	};
 	const tableRef = useRef();
+	const loadComponent = (component, callback) => {
+		let element = null;
+		component = component.toLowerCase();
+		Object.keys(ComponentMap).forEach(k => {
+			if (k.toLowerCase() === component.toLowerCase()) {
+				element = k;
+			}
+		});
+		if (!element) return;
+		if (ComponentMap[element] === null) {
+			import('../components/'+component+'/'+component).then((m) => {
+				ComponentMap[element] = m.default;
+				callback();
+			});
+		} else {
+			callback();
+		}
+	}
 
 	useEffect(() => {
-		// Auto update is only available when all modal are closed.
-		if (!runner && !desktop && !procMgr && !explorer && !generate && !terminal) {
+		// auto update is only available when all modal are closed.
+		if (!execute && !desktop && !procMgr && !explorer && !generate && !terminal) {
 			let id = setInterval(getData, 3000);
 			return () => {
 				clearInterval(id);
 			};
 		}
-	}, [runner, desktop, procMgr, explorer, generate, terminal]);
+	}, [execute, desktop, procMgr, explorer, generate, terminal]);
 
 	function getInitColumnsState() {
 		let data = localStorage.getItem(`columnsState`);
@@ -293,7 +292,7 @@ function overview(props) {
 	}
 	function renderOperation(device) {
 		let menus = [
-			{key: 'run', name: i18n.t('OVERVIEW.RUN')},
+			{key: 'execute', name: i18n.t('OVERVIEW.EXECUTE')},
 			{key: 'desktop', name: i18n.t('OVERVIEW.DESKTOP')},
 			{key: 'screenshot', name: i18n.t('OVERVIEW.SCREENSHOT')},
 			{key: 'lock', name: i18n.t('OVERVIEW.LOCK')},
@@ -305,29 +304,33 @@ function overview(props) {
 			{key: 'offline', name: i18n.t('OVERVIEW.OFFLINE')},
 		];
 		return [
-			<a key='terminal' onClick={setTerminal.bind(null, device)}>{i18n.t('OVERVIEW.TERMINAL')}</a>,
-			<a key='procmgr' onClick={setProcMgr.bind(null, device.id)}>{i18n.t('OVERVIEW.PROC_MANAGER')}</a>,
-			<a key='explorer' onClick={() => {
-				setExplorer(device.id);
-				setIsWindows(device.os === 'windows');
-			}}>
-				{i18n.t('OVERVIEW.EXPLORER')}
-			</a>,
+			<a key='terminal' onClick={() => onMenuClick('terminal', device)}>{i18n.t('OVERVIEW.TERMINAL')}</a>,
+			<a key='explorer' onClick={() => onMenuClick('explorer', device)}>{i18n.t('OVERVIEW.EXPLORER')}</a>,
+			<a key='procmgr' onClick={() => onMenuClick('procmgr', device)}>{i18n.t('OVERVIEW.PROC_MANAGER')}</a>,
 			<TableDropdown
 				key='more'
-				onSelect={key => callDevice(key, device)}
+				onSelect={key => onMenuClick(key, device)}
 				menus={menus}
 			/>,
 		]
 	}
 
-	function callDevice(act, device) {
-		if (act === 'run') {
-			setRunner(device);
-			return;
-		}
-		if (act === 'desktop') {
-			setDesktop(device);
+	function onMenuClick(act, value) {
+		const device = value;
+		let hooksMap = {
+			terminal: setTerminal,
+			explorer: setExplorer,
+			generate: setGenerate,
+			procmgr: setProcMgr,
+			execute: setExecute,
+			desktop: setDesktop,
+		};
+		if (hooksMap[act]) {
+			setLoading(true);
+			loadComponent(act, () => {
+				hooksMap[act](device);
+				setLoading(false);
+			});
 			return;
 		}
 		if (act === 'screenshot') {
@@ -360,7 +363,7 @@ function overview(props) {
 
 	function toolBar() {
 		return (
-			<Button type='primary' onClick={setGenerate.bind(null, true)}>{i18n.t('OVERVIEW.GENERATE')}</Button>
+			<Button type='primary' onClick={() => onMenuClick('generate', true)}>{i18n.t('OVERVIEW.GENERATE')}</Button>
 		)
 	}
 
@@ -413,7 +416,7 @@ function overview(props) {
 		<>
 			<Image
 				preview={{
-					visible: screenBlob,
+					visible: !!screenBlob,
 					src: screenBlob,
 					onVisibleChange: () => {
 						URL.revokeObjectURL(screenBlob);
@@ -421,36 +424,53 @@ function overview(props) {
 					}
 				}}
 			/>
-			<Generate
-				visible={generate}
-				onVisibleChange={setGenerate}
-			/>
-			<Explorer
-				isWindows={isWindows}
-				visible={explorer}
-				device={explorer}
-				onCancel={setExplorer.bind(null, false)}
-			/>
-			<ProcMgr
-				visible={procMgr}
-				device={procMgr}
-				onCancel={setProcMgr.bind(null, false)}
-			/>
-			<Runner
-				visible={runner}
-				device={runner}
-				onCancel={setRunner.bind(null, false)}
-			/>
-			<Desktop
-				visible={desktop}
-				device={desktop}
-				onCancel={setDesktop.bind(null, false)}
-			/>
-			<Terminal
-				visible={terminal}
-				device={terminal}
-				onCancel={setTerminal.bind(null, false)}
-			/>
+			{
+				ComponentMap.Generate &&
+				<ComponentMap.Generate
+					visible={generate}
+					onVisibleChange={setGenerate}
+				/>
+			}
+			{
+				ComponentMap.Execute &&
+				<ComponentMap.Execute
+					visible={execute}
+					device={execute}
+					onCancel={setExecute.bind(null, false)}
+				/>
+			}
+			{
+				ComponentMap.Explorer &&
+				<ComponentMap.Explorer
+					open={explorer}
+					device={explorer}
+					onCancel={setExplorer.bind(null, false)}
+				/>
+			}
+			{
+				ComponentMap.ProcMgr &&
+				<ComponentMap.ProcMgr
+					open={procMgr}
+					device={procMgr}
+					onCancel={setProcMgr.bind(null, false)}
+				/>
+			}
+			{
+				ComponentMap.Desktop &&
+				<ComponentMap.Desktop
+					open={desktop}
+					device={desktop}
+					onCancel={setDesktop.bind(null, false)}
+				/>
+			}
+			{
+				ComponentMap.Terminal &&
+				<ComponentMap.Terminal
+					open={terminal}
+					device={terminal}
+					onCancel={setTerminal.bind(null, false)}
+				/>
+			}
 			<ProTable
 				scroll={{
 					x: 'max-content',
@@ -464,6 +484,8 @@ function overview(props) {
 					value: columnsState,
 					onChange: saveColumnsState
 				}}
+				onLoadingChange={setLoading}
+				loading={loading}
 				request={getData}
 				pagination={false}
 				actionRef={tableRef}
@@ -472,6 +494,27 @@ function overview(props) {
 				onDataSourceChange={setDataSource}
 			/>
 		</>
+	);
+}
+function UsageBar(props) {
+	let {usage} = props;
+	usage = usage || 0;
+	usage = Math.round(usage * 100) / 100;
+
+	return (
+		<Tooltip
+			title={props.title??`${usage}%`}
+			overlayInnerStyle={{
+				whiteSpace: 'nowrap',
+				wordBreak: 'keep-all',
+				maxWidth: '300px',
+			}}
+			overlayStyle={{
+				maxWidth: '300px',
+			}}
+		>
+			<Progress percent={usage} showInfo={false} strokeWidth={12} trailColor='#FFECFF'/>
+		</Tooltip>
 	);
 }